chore: use date string for in DB sorting.
Description: This major changes brings optimization and reduce cpu and memory usages. As well as using build-in mongodb tools for sorting, pagination, aggregations for better performance. Which reduce loads on server as well as client side application. Added two migration files for converting existing createTime, verifyTime and updateTime field's string to mongoDB supported Date objects.

Why?: MongoDB supports sorting on date fields if they are stored as Date objects. If dates is stored as strings, MongoDB will sort them lexicographically, which may not give the expected results if the date strings are not in a format that sorts correctly lexicographically (like ISO 8601).

Previousely createTime field is stored as a string in the format of new Date().toLocaleString(). This format does not sort correctly lexicographically because it includes components like the name of the month.